Rename ./mutator/agentconfig to ./agentmap
