+++ cmd/traffic/cmd/manager/internal/agentmap/capsbase26.go
@@ -1,4 +1,4 @@
-package agentconfig
+package agentmap
 
 // CapsBase26 converts the given number into base 26 represented using the letters 'A' to 'Z'
 func CapsBase26(v uint64) string {
