+++ cmd/traffic/cmd/manager/internal/agentmap/capsbase26_test.go
@@ -1,4 +1,4 @@
-package agentconfig
+package agentmap
 
 import (
 	"math"
