+++ cmd/traffic/cmd/manager/internal/agentmap/discorvery.go
@@ -1,4 +1,4 @@
-package agentconfig
+package agentmap
 
 import (
 	"context"
