+++ cmd/traffic/cmd/manager/internal/agentmap/watcher.go
@@ -1,4 +1,4 @@
-package agentconfig
+package agentmap
 
 import (
 	"bytes"
